+++ hata/discord/client/client.py
@@ -39,7 +39,8 @@ from ..events.intent import IntentFlag
 from ..events.core import register_client, unregister_client
 from ..audit_logs import AuditLog, AuditLogIterator, AuditLogEvent
 from ..invite import Invite
-from ..message import Message, MessageRepr, MessageReference, Attachment, Sticker, MessageFlag
+from ..message import Message, MessageRepr, MessageReference, Attachment, Sticker, MessageFlag, \
+    try_resolve_interaction_message
 from ..oauth2 import Connection, parse_locale, DEFAULT_LOCALE, OA2Access, UserOA2, Achievement
 from ..exceptions import DiscordException, DiscordGatewayException, ERROR_CODES, InvalidToken
 from ..core import CLIENTS, KOKORO, GUILDS, DISCOVERY_CATEGORIES, EULAS, CHANNELS, EMOJIS, APPLICATIONS, ROLES, \
@@ -13721,8 +13722,9 @@ class Client(ClientUserPBase):
             message_data = await self.http.interaction_followup_message_create(application_id, interaction.id,
                 interaction.token, message_data)
         
-        return interaction.channel._create_new_message(message_data)
-    
+        message = interaction.channel._create_new_message(message_data)
+        try_resolve_interaction_message(message, interaction)
+        return message
     
     async def interaction_followup_message_edit(self, interaction, message, content=..., *, embed=..., file=None,
             allowed_mentions=...,):